refactor: organizing sections in one folder
