+++ src/pages/sections/AcademicLife/index.tsx
