+++ src/pages/sections/HardSkills/content/languages.ts
