+++ src/pages/sections/HardSkills/content/tools.ts
@@ -1,117 +1,117 @@
 export const tools = [
   {
     title: "Node.js",
-    image: require("../../../assets/tools/nodejs.jpeg"),
+    image: require("../../../../assets/tools/nodejs.jpeg"),
     alt: "Node.js image.",
   },
   {
     title: "Express.js",
-    image: require("../../../assets/tools/expressjs.jpeg"),
+    image: require("../../../../assets/tools/expressjs.jpeg"),
     alt: "Express.js image.",
   },
   {
     title: "Nest.js",
-    image: require("../../../assets/tools/nestjs.jpeg"),
+    image: require("../../../../assets/tools/nestjs.jpeg"),
     alt: "Nest.js image.",
   },
   {
     title: "Fastify",
-    image: require("../../../assets/tools/fastify.jpeg"),
+    image: require("../../../../assets/tools/fastify.jpeg"),
     alt: "Fastify.js image.",
   },
   {
     title: "Web3.js",
-    image: require("../../../assets/tools/web3js.jpg"),
+    image: require("../../../../assets/tools/web3js.jpg"),
     alt: "Web3.js image.",
   },
   {
     title: "Ethers.js",
-    image: require("../../../assets/tools/ethersjs.jpeg"),
+    image: require("../../../../assets/tools/ethersjs.jpeg"),
     alt: "Ethers.js image.",
   },
   {
     title: "Hardhat",
-    image: require("../../../assets/tools/hardhat.jpeg"),
+    image: require("../../../../assets/tools/hardhat.jpeg"),
     alt: "Hardhat image.",
   },
   {
     title: "Truffle",
-    image: require("../../../assets/tools/truffle.jpeg"),
+    image: require("../../../../assets/tools/truffle.jpeg"),
     alt: "Truffle image.",
   },
   {
     title: "Ethereum",
-    image: require("../../../assets/tools/ethereum.jpeg"),
+    image: require("../../../../assets/tools/ethereum.jpeg"),
     alt: "Ethereum image.",
   },
   {
     title: "Binance Smart Chain",
-    image: require("../../../assets/tools/bsc.jpeg"),
+    image: require("../../../../assets/tools/bsc.jpeg"),
     alt: "Binance Smart Chain image.",
   },
   {
     title: "TRON",
-    image: require("../../../assets/tools/tron.jpeg"),
+    image: require("../../../../assets/tools/tron.jpeg"),
     alt: "Tron image.",
   },
   {
     title: "Polygon",
-    image: require("../../../assets/tools/polygon.jpeg"),
+    image: require("../../../../assets/tools/polygon.jpeg"),
     alt: "Polygon image.",
   },
   {
     title: "MongoDB",
-    image: require("../../../assets/tools/mongodb.jpeg"),
+    image: require("../../../../assets/tools/mongodb.jpeg"),
     alt: "MongoDB image.",
   },
   {
     title: "MySQL",
-    image: require("../../../assets/tools/mysql.jpeg"),
+    image: require("../../../../assets/tools/mysql.jpeg"),
     alt: "MySQL image.",
   },
   {
     title: "PostgreSQL",
-    image: require("../../../assets/tools/postgres.jpeg"),
+    image: require("../../../../assets/tools/postgres.jpeg"),
     alt: "PostgreSQL image.",
   },
   {
     title: "Redis",
-    image: require("../../../assets/tools/redis.jpeg"),
+    image: require("../../../../assets/tools/redis.jpeg"),
     alt: "Redis image.",
   },
   // {
   //   title: "React.js",
-  //   image: require("../../../assets/tools/reactjs.jpeg"),
+  //   image: require("../../../../assets/tools/reactjs.jpeg"),
   //   alt: "React.js image.",
   // },
   // {
   //   title: "Styled Components",
-  //   image: require("../../../assets/tools/styled-components.jpeg"),
+  //   image: require("../../../../assets/tools/styled-components.jpeg"),
   //   alt: "Styled Components image.",
   // },
   // {
   //   title: "Figma",
-  //   image: require("../../../assets/tools/figma.jpeg"),
+  //   image: require("../../../../assets/tools/figma.jpeg"),
   //   alt: "Figma image.",
   // },
   // {
   //   title: "Remix IDE",
-  //   image: require("../../../assets/tools/remixide.jpeg"),
+  //   image: require("../../../../assets/tools/remixide.jpeg"),
   //   alt: "Remix IDE image.",
   // },
   // {
   //   title: "Git",
-  //   image: require("../../../assets/tools/git.jpeg"),
+  //   image: require("../../../../assets/tools/git.jpeg"),
   //   alt: "Git image.",
   // },
   // {
   //   title: "Digital Ocean",
-  //   image: require("../../../assets/tools/digitalocean.jpeg"),
+  //   image: require("../../../../assets/tools/digitalocean.jpeg"),
   //   alt: "Digital Ocean image.",
   // },
   // {
   //   title: "Google Cloud",
-  //   image: require("../../../assets/tools/googlecloud.jpeg"),
+  //   image: require("../../../../assets/tools/googlecloud.jpeg"),
   //   alt: "Google Cloud image.",
   // },
 ];
