+++ src/pages/sections/HardSkills/index.tsx
