+++ src/pages/sections/HardSkills/styles.tsx
