+++ src/pages/sections/Quote/content/quotes.ts
