+++ src/pages/sections/Quote/styles.tsx
@@ -44,7 +44,7 @@ export const VerticalLine = styled.div(
 );
 
 export const QuoteImg = styled.img.attrs({
-  src: require("../../assets/quote.svg").default,
+  src: require("../../../assets/quote.svg").default,
   alt: "A quote icon.",
 })`
   position: absolute;
@@ -58,7 +58,7 @@ export const QuoteImg = styled.img.attrs({
 `;
 
 export const QuoteImgReverse = styled.img.attrs({
-  src: require("../../assets/quote.svg").default,
+  src: require("../../../assets/quote.svg").default,
   alt: "A quote icon.",
 })`
   position: absolute;
