+++ src/pages/sections/SoftSkills/content/softskills.ts
