+++ src/pages/sections/SoftSkills/index.tsx
