+++ src/pages/sections/SoftSkills/styles.tsx
